Arithmetic: outputting a warning message on overflows
Until this commit, arithmetic provides no warning when an arithmetic
operation over/underflows. Users would have to either live with overflows
or manually check for them, neither of which options are ideal.

With this commit, arithmetic provides warnings whenever an overflow
occurs on an addition, subtraction, multiplication, or division. The
overflow check can be turned on and off by the user. If turned on,
overflow warning messages are not suppressed by the -q option.

This fixes task #16114: https://savannah.gnu.org/task/?16114