Added run-time option to Arithmetic for doing overflow check or not
Until now, we had to manually set the overflow flag. Also, the warning
message was too minimal (not too effective for an astronomer who confronts
this issue for the first time), and the message was executed within each
low-level function (making it hard to maintain).

With this commit, a run-time option has been added for this. Also, the four
affected functions now return an 'int' (which shows if any overflow
occurred or not), and a complete warning message has been added in the
higher-level function (in one place).