Library (error.h): high-level codes for errors
Until Now, the error mechnaism only retained information about the library
and function which caused the error. However, errors can be classified into
higher-level "type"s (which are necessary in interfaces like Python).

With this commit, A new 'enum' structure has been added in
'lib/gnuastro/error.h' which is primarily based on the C library error
codes. If any useful exception identifiers are available in Python, they
have also been included (and clearly identified to be from Python).

This Gnuastro-wide error code will be used from now on for each error, not
a library-specific code. The low-level details are too specific to be coded
outside of one function and the human-readable message is enough (as in the
GNU C library which is orders of magnitude larger than Gnuastro's library).

The Python exceptions were collected by Labeeb Asari.