All programs: new --list-error-codes to print error codes and info
Until now, when an error occurred Gnuastro's programs would simply return
'EXIT_FAILURE' (1). This would not let the caller easily identify the
source of the problem and be prepared for it!

With this commit, the preparations have been done for all the programs to
return the same error codes as we define for the libraries to the shell. To
help the users interpret those returned integers, we also a new common
option called '--list-error-codes' has been added to all the programs.