All programs: --list-error-codes includes error name and category
Until now, when printing the list of errors with '--list-error-codes', the
programs would only print the raw integer and description. But that is too
annoying/hard to check for higher-level things (for example if you already
know the standard C library error names or if you only want to see the
category of errors related to files).

With this commit, the printed table contains two extra columns: the
standard name of the error and the category. The name can be very useful
for the advanced users of Gnuastro because that will not be changed, but
the integer may change (if we add/remove a new code).

In the process (following Labib Asari's suggestion in the Gnuastro
developers meeting), it was decided that functions which do multiple
breaking error checks should stop on the first occasion, and we should try
to avoid having multiple breaking errors in one error list. This new
strategy has now been implemented in 'cosmology_sanity_check'.