Library (progcrash.h): to be compiled by each program separately
Until now, we had the simple 'lib/gnuastro/errorinprogram.h' for a common
set of full function definitions to crash programs in case of
errors. However, this would result in the same function definition in two
object files; causing a linking error in the programs.

With this commit, this file has been taken to the
'gnuastro/lib/progcrash.c', but it is not added to the library source
(since we don't want to have any 'exit' commands in the library). All the
separate programs that use it should not include this file in their
separate Makefiles.