WIP: Segment: first draft of accounting for azimuth in clump boundary
Until now, a small clump on the wing of a larger source of light would
eat-up the diffuse flux from the larger object. For more, see

https://savannah.gnu.org/bugs/index.php?62702#comment1

With this clump, the per-pixel checking of the azimuthal angle is added to
draw a plot and check the hypothesis that the azimuthal angle can be used
to find the boundary of such clumps.

WIP: using the counts-per-second correction for S/N

Segment: WIP: corrected min1_max0

Segment: WIP: keepmaxnearriver to nomaxnearriver