Deconvolve: New program implementation
Until now, there was no deconvolution program in Gnuastro. Deconvolution is
a needed functionality for improving astonomical images and decreasing of
the blurred observations.

With this commit, a new deconvolution program has been added to Gnuastro
for this purpose. To achieve this, several low-level libraries have been
added to provide the foundation for the program, but can also be reused in
future developments. These libraries are:

  - complex.h: Data structures and operations for complex number arrays.
  - fft.h: Fast Fourier Transform (FFT) algorithms (direct and reverse
    way).
  - wavelet.h: Implements wavelet transforms that allow for
    multi-resolution analysis and better handling of noise in deconvolution
    processes.
  - deconvolve.h: Contains the core functionality for performing
    deconvolution on images using various algorithms.

The new program adds a user-friendly way to work with the deconvolution
library via command line. It offers 4 algorithms:

   - Direct: A simple method that attempts to invert the blurring process
     directly, suitable for cases where the blur kernel is well-known.
   - Tikhonov: A regularized deconvolution method that helps deal with
     ill-posed problems by adding a penalty term to the solution.
   - RL: Richardson-Lucy deconvolution, an iterative method that improves
     the quality of the image by progressively refining the estimate of the
     original image.
   - AWMLE: An advanced wavelet-based method that uses a Maximum Likelihood
     Estimation approach for more robust deconvolution in the presence of
     noise.