WIP: read message
The single tiles are chosen nicely! We are now in the phase of removing
outliers.

Histogram and cumulative frequency plots are no good!

What seems to be good is to start from the smallest value tiles and come
up. For each tile, check the neighbors and if time the dispersion between
the neighboring values gets too much, we should stop and reject all other
tiles.