Convolve: Fixed kernel offset
Until now, the kernel was not placed correctly after the padding. This
issue was creating an offset that affected the final image.

With this commit the kernel is now placed correctly. The central pixel
must to be placed in the (0,0) to "break" the kernel and split it
along the 4 corners. Now, with the "--checkfreqsteps" option set, the
fourier domain kernel no longer has diagonal lines, and makekernel
doesn't need to fix at the end of the operations.